refactor: mv ChannelInList
