+++ src/components/ChannelInList.tsx
